feat(api): update leader ranks to reflect api change of totals seconds from int to double. Update readme

+++ models/leaders_rank.go
@@ -188,7 +188,7 @@ type LeadersRankRunningTotalLanguagesItems0 struct {
 	Name string `json:"name,omitempty"`
 
 	// total seconds
-	TotalSeconds int64 `json:"total_seconds,omitempty"`
+	TotalSeconds float64 `json:"total_seconds,omitempty"`
 }
 
 // Validate validates this leaders rank running total languages items0
